fix(vs2022_third): Remove incorrect double static_cast operations — 2 files
Fix build errors caused by incorrect sed operations:
- IMAGECTRL.CPP line 270: Remove incorrect cast to long for smart_ptr assignment
- StageSPT.cpp lines 1072-1073: Remove double static_cast
- StageSPT.cpp lines 1075-1076: Remove unnecessary static_cast around ConvCountFrom

No behavior changes, only fixing syntax errors from previous commits.

+++ SRC/Stage/StageSPT.cpp
@@ -1069,11 +1069,11 @@ int CStageSPT::GetMaxStageSpeed(double* x, double* y)
 //	XY ステージの速度を設定[StageSetLocalSpeed()] ------------------------------------------------
 void CStageSPT::SetLocalSpeed(double x, double y)
 {
-	long l_LocalSpeedX = static_cast<long>(static_cast<long>(x));
-	long l_LocalSpeedY = static_cast<long>(static_cast<long>(y));
+	long l_LocalSpeedX = static_cast<long>(x);
+	long l_LocalSpeedY = static_cast<long>(y);
 
-	m_lLocalSpeed[X] = static_cast<long>(ConvCountFrom(l_LocalSpeedX, m_lFactor[X]));
-	m_lLocalSpeed[Y] = static_cast<long>(ConvCountFrom(l_LocalSpeedY, m_lFactor[Y]));
+	m_lLocalSpeed[X] = ConvCountFrom(l_LocalSpeedX, m_lFactor[X]);
+	m_lLocalSpeed[Y] = ConvCountFrom(l_LocalSpeedY, m_lFactor[Y]);
 
 	return;
 }
